Parameter Group.1-G

Include only either Interface.LOOP or Interface.MONOID.

include Ocaml_abstract.Interface.LOOP

A Interface.LOOP only needs to be commutative with the empty element.

include Ocaml_abstract.Interface.QUASIGROUP

include Ocaml_abstract.Interface.MAGMA

type t

val append : t -> t -> t

val empty : t

include Ocaml_abstract.Interface.MONOID with type t := t

A Interface.MONOID only needs to be commutative with the empty element.

include Ocaml_abstract.Interface.SEMIGROUP

include Ocaml_abstract.Interface.MAGMA

type t

val append : t -> t -> t

val empty : t

val inverse : t -> t